test(reestablish): move commit diff tests into dedicated module
